feat: Create CodeXPUIManager to move the logic of displaying the gained XP label from CodeXPEventListener
